chore: refactor shared code between services
