+++ gapic/templates/%namespace/%name_%version/%sub/_compat/__init__.py.j2
@@ -1,0 +1,1 @@
+{% extends '_base.py.j2' %}
